requirements.txt
# RHYMIA 챗봇 - OpenAI + LangChain 기반
# 설치: pip install -r requirements.txt

# OpenAI 공식 Python 클라이언트
openai>=1.0.0

# LangChain 코어 (프롬프트, 메시지 등)
langchain-core>=0.3.0

# LangChain의 OpenAI 채팅 모델 (ChatOpenAI)
langchain-openai>=0.2.0

# .env 파일에서 환경 변수 로드
python-dotenv>=1.0.0
